testing and refactoring

+++ src/main/java/infrastructure/system/Leader.java
@@ -2,5 +2,5 @@ package infrastructure.system;
 
 import java.net.InetAddress;
 
-public record Leader(InetAddress leaderIp, short leaderPort) {
+public record Leader(InetAddress leaderIp, int leaderPort) {
 }
